pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "qesdk"
version = "0.0.6"
authors = [
  { name="Scott Zhang", email="scott2011@qq.com" },
]
description = "Quantease SDK for quants"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  'msgpack_python>=0.5.6',
  'nest_asyncio>=1.5.5',
  'numpy>=1.23.2',
  'pandas>=1.3.5',
  'thriftpy2>=0.4.14',
]
[project.urls]
"Homepage" = "https://github.com/quantease/qesdk"
"Bug Tracker" = "https://github.com/quantease/qesdk/issues"

